+++ index.js
@@ -1,0 +1,3 @@
+const { checkVersions } = require('./src/utils');
+checkVersions();
+module.exports = require('./src');